% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhsR-package.R
\docType{package}
\name{qhsR-package}
\alias{qhsR}
\alias{qhsR-package}
\title{qhsR: R Functions for the CCF QHS Department}
\description{
A miscellaneous collection of R functions for use by members of the CCF QHS department.
}
\author{
\strong{Maintainer}: Emily C. Zabor \email{zabore2@ccf.org} (\href{https://orcid.org/0000-0002-1402-4498}{ORCID})

}
\keyword{internal}
